third_party/blink/web_tests/virtual/document-transition-wide-gamut/document-transition/capture-callback-exception-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: foo is not defined
CONSOLE ERROR: Uncaught (in promise) AbortError: Aborted due to failure in DocumentTransitionCallback
CONSOLE ERROR: Uncaught (in promise) AbortError: Aborted due to failure in DocumentTransitionCallback
